examples/example-simple/requirements/reqs_test.txt
#
# This file is autogenerated by pip-compile
# Make changes in reqs_test.in, then run this to update:
#
#    pip-compile reqs_test.in
#
coverage==4.0.3
cssselect==0.9.1          # via pyquery
factory-boy==2.6.1
fake-factory==0.5.7       # via factory-boy
inflection==0.3.1         # via pytest-factoryboy
lxml==3.6.0               # via pyquery
py==1.4.31                # via pytest
pyquery==1.2.11
pytest-cov==2.2.1
pytest-django==2.9.1
pytest-factoryboy==1.1.5
pytest==2.9.1
python-dateutil==2.5.2    # via fake-factory
six==1.10.0               # via fake-factory, python-dateutil
